Lots of cool layout changes!

+++ app/src/main/java/app/mblackman/orderfulfillment/data/domain/Product.kt
@@ -7,7 +7,7 @@ data class Product(
     val name: String,
     val description: String,
     val imageUrl: String?,
-    val price: Float?
+    val price: Double?
 )
 
 /**
